pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "costetl"
version = "0.1.0"
description = "Excel Cost Report ETL System"
authors = [{name = "Data Engineering Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "SQLAlchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dateutil>=2.8.0",
    "rapidfuzz>=3.0.0",
    "PyYAML>=6.0",
    "click>=8.0.0",
    "great-expectations>=0.17.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "types-PyYAML",
    "types-python-dateutil",
]

[project.scripts]
costetl = "costetl.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=costetl --cov-report=term-missing"